requirements.txt
streamlit==1.36.0
pandas==1.5.3
azure-search-documents==11.4.0
tabulate
plotly
sentence_transformers
openai
jupyter
ipywidgets